prueba

+++ .ccls-cache/@home@runner@Proyecto-SuperStore/Funciones@adjacentListsDestroyer.c
@@ -6,7 +6,7 @@ void destroyAllAdjacentListsProductos(HashMap* mapaProductos)
     tipoProducto* productoCurrent = parAux->value;
     if (productoCurrent == NULL)
         return;
-    while (productoCurrent != NULL)
+    while (parAux != NULL)
     {
         productoCurrent = parAux->value;
         if (productoCurrent->supermercados != NULL)
@@ -15,17 +15,18 @@ void destroyAllAdjacentListsProductos(HashMap* mapaProductos)
     }
 }
 
-/*
+
 void destroyAllAdjacentListsSupermercados(HashMap* mapaSupermercados)
 {
     Pair* parAux = firstMap(mapaSupermercados);
     tipoSupermercado* supermercadoCurrent = parAux->value;
     if (supermercadoCurrent == NULL)
         return;
-    while (supermercadoCurrent != NULL)
+    while (parAux != NULL)
     {
         supermercadoCurrent = parAux->value;
-        destroyList(supermercadoCurrent->productos);
+        if (supermercadoCurrent->productos != NULL)
+            destroyList(supermercadoCurrent->productos);
         parAux = nextMap(mapaSupermercados);
     }
 }
@@ -36,18 +37,40 @@ void destroyAllAdjacentListsCategorias(HashMap* mapaCategorias)
     tipoSupermercado* categoriaCurrent = parAux->value;
     if (categoriaCurrent == NULL)
         return;
-    while (categoriaCurrent != NULL)
+    while (parAux != NULL)
     {
         categoriaCurrent = parAux->value;
-        destroyList(categoriaCurrent->productos);
+        if (categoriaCurrent->productos != NULL)
+            destroyList(categoriaCurrent->productos);
         parAux = nextMap(mapaCategorias);
     }
 }
 
-*/
-//void destroyAllAdjacentListsBTree(BTree* arbolB)
-//{
-    
-//}
 
+void destroyAllAdjacentListsBTree(BTreeNode *node)
+{
+    if (node == NULL)
+        return;
+
+    // Si el nodo no es una hoja, llamar a BTreeDestroy en cada hijo
+    if (!isLeaf(node)) {
+        for (int i = 0; node->children[i] != NULL; i++) {
+            if (node->children[i] != NULL)
+                destroyAllAdjacentListsBTree(node->children[i]);
+        }
+    }
+
+    // Liberar los recursos para las claves y los hijos
+    if (node->keys != NULL) {
+        free(node->keys);
+        node->keys = NULL;
+    }
+    if (node->children != NULL) {
+        free(node->children);
+        node->children = NULL;
+    }
+
+    // Finalmente, liberar el nodo mismo
+    free(node);
+}
